Struct bip_metainfo::MetainfoBuilder

pub struct MetainfoBuilder<'a> { /* fields omitted */ }

Builder for generating a torrent file from some accessor.

Methods

impl<'a> MetainfoBuilder<'a>

fn new() -> MetainfoBuilder<'a>

Create a new MetainfoBuilder with some default values set.

fn set_main_tracker(self, tracker_url: &'a str) -> MetainfoBuilder<'a>

Set the main tracker that this torrent file points to.

fn set_creation_date(self, secs_epoch: i64) -> MetainfoBuilder<'a>

Set the creation date for the torrent.

Defaults to the current time when the builder was created.

fn set_comment(self, comment: &'a str) -> MetainfoBuilder<'a>

Set a comment for the torrent file.

fn set_created_by(self, created_by: &'a str) -> MetainfoBuilder<'a>

Set the created by for the torrent file.

fn set_private_flag(self, is_private: bool) -> MetainfoBuilder<'a>

Sets the private flag for the torrent file.

fn set_piece_length(self, piece_length: PieceLength) -> MetainfoBuilder<'a>

Sets the piece length for the torrent file.

fn build_as_bytes<A, C>(
self,
threads: usize,
accessor: A,
progress: C
) -> ParseResult<Vec<u8>> where
A: IntoAccessor,
C: FnMut(f64) + Send + 'static,

Build the metainfo file from the given accessor and the number of worker threads.

Worker threads are responsible for CPU bound tasks so if IO access is slow, increasing the number of workers may not be beneficial. This method WILL block until it completes.

Returns a list of bytes that make up the complete metainfo file.

Panics if threads is equal to zero.